feat: basic string adapter
