+++ packages/amplify-graphql-schema-generator/src/datasource-adapter/index.ts
@@ -3,4 +3,3 @@ export * from './string-datasource-adapter';
 export * from './mysql-datasource-adapter';
 export * from './pg-datasource-adapter';
 export * from './mysql-string-datasource-adapter';
-export * from './pg-string-datasource-adapter';
